pyproject.toml
[project]
name = "svj-sprava"
version = "0.1.0"
description = "SVJ automation: per rollam voting, tax distribution, owner sync"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "jinja2>=3.1.4",
    "python-multipart>=0.0.12",
    "sqlalchemy>=2.0.36",
    "pydantic-settings>=2.6.0",
    "openpyxl>=3.1.5",
    "python-docx>=1.1.2",
    "docxtpl>=0.18.0",
    "pdfplumber>=0.11.4",
    "pytesseract>=0.3.13",
    "Pillow>=11.0.0",
    "unidecode>=1.3.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "httpx>=0.28.0",
]
